chore: apply suggestions and merge benchmark classes
